Improve platform and Coap abstraction
Refactor lwm2m to make it easier to support other client platforms in the future:
- Move linux and contiki specific support to platform specific util modules.
- Move Coap specific processing to coap abstraction layer.
