+++ core/src/common/coap_abstraction.h
@@ -91,6 +91,8 @@ void coap_SendNotify(AddressType * addr, const char * path, const char * token, 
 void coap_SetContext(void * ctxt);
 void coap_SetRequestHandler(RequestHandler handler);
 
+int coap_WaitMessage(int timeout, int fd);
+
 int coap_ResolveAddressByURI(unsigned char * address, AddressType * addr);
 
 int coap_RegisterUri(const char * uri);
